Drop V6 implementation.
