+++ cryptomator/src/main/java/ch/cyberduck/core/cryptomator/impl/CryptoDirectoryV7Provider.java
@@ -15,37 +15,58 @@ package ch.cyberduck.core.cryptomator.impl;
  * GNU General Public License for more details.
  */
 
+import ch.cyberduck.core.AbstractPath;
+import ch.cyberduck.core.CacheReference;
 import ch.cyberduck.core.Path;
+import ch.cyberduck.core.PathAttributes;
 import ch.cyberduck.core.RandomStringService;
 import ch.cyberduck.core.Session;
+import ch.cyberduck.core.SimplePathPredicate;
 import ch.cyberduck.core.UUIDRandomStringService;
+import ch.cyberduck.core.cache.LRUCache;
 import ch.cyberduck.core.cryptomator.AbstractVault;
 import ch.cyberduck.core.cryptomator.ContentReader;
+import ch.cyberduck.core.cryptomator.CryptoDirectory;
 import ch.cyberduck.core.cryptomator.CryptoFilename;
 import ch.cyberduck.core.cryptomator.CryptorCache;
 import ch.cyberduck.core.exception.BackgroundException;
 import ch.cyberduck.core.exception.NotfoundException;
+import ch.cyberduck.core.preferences.PreferencesFactory;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.nio.charset.StandardCharsets;
 import java.util.EnumSet;
+import java.util.concurrent.locks.ReadWriteLock;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 import com.google.common.io.BaseEncoding;
 
-public class CryptoDirectoryV7Provider extends CryptoDirectoryV6Provider {
+public class CryptoDirectoryV7Provider implements CryptoDirectory {
     private static final Logger log = LogManager.getLogger(CryptoDirectoryV7Provider.class);
 
+    private static final String DATA_DIR_NAME = "d";
+
+    public static final byte[] ROOT_DIR_ID = new byte[0];
+
+    private final AbstractVault vault;
+    private final Path dataRoot;
+    private final Path home;
     private final CryptoFilename filenameProvider;
     private final CryptorCache filenameCryptor;
-    private final AbstractVault vault;
 
     private final RandomStringService random
             = new UUIDRandomStringService();
 
+    protected final ReadWriteLock lock = new ReentrantReadWriteLock();
+    protected final LRUCache<CacheReference<Path>, byte[]> cache = LRUCache.build(
+            PreferencesFactory.get().getInteger("cryptomator.cache.size"));
+
+
     public CryptoDirectoryV7Provider(final AbstractVault vault, final CryptoFilename filenameProvider, final CryptorCache filenameCryptor) {
-        super(vault.getHome(), filenameProvider, filenameCryptor);
+        this.home = vault.getHome();
+        this.dataRoot = new Path(vault.getHome(), DATA_DIR_NAME, vault.getHome().getType());
         this.filenameProvider = filenameProvider;
         this.filenameCryptor = filenameCryptor;
         this.vault = vault;
@@ -56,6 +77,101 @@ public class CryptoDirectoryV7Provider extends CryptoDirectoryV6Provider {
         final String ciphertextName = filenameCryptor.encryptFilename(BaseEncoding.base64Url(), filename, this.getOrCreateDirectoryId(session, parent)) + vault.getRegularFileExtension();
         log.debug("Encrypted filename {} to {}", filename, ciphertextName);
         return filenameProvider.deflate(session, ciphertextName);
+    }
+
+    @Override
+    public Path toEncrypted(final Session<?> session, final Path directory) throws BackgroundException {
+        if(!directory.isDirectory()) {
+            throw new NotfoundException(directory.getAbsolute());
+        }
+        if(new SimplePathPredicate(directory).test(home) || directory.isChild(home)) {
+            final PathAttributes attributes = new PathAttributes(directory.attributes());
+            // The root of the vault is a different target directory and file ids always correspond to the metadata file
+            attributes.setVersionId(null);
+            attributes.setFileId(null);
+            // Remember random directory id for use in vault
+            final byte[] id = this.getOrCreateDirectoryId(session, directory);
+            log.debug("Use directory ID '{}' for folder {}", id, directory);
+            attributes.setDirectoryId(id);
+            attributes.setDecrypted(directory);
+            final String directoryIdHash = filenameCryptor.hashDirectoryId(id);
+            // Intermediate directory
+            final Path intermediate = new Path(dataRoot, directoryIdHash.substring(0, 2), dataRoot.getType());
+            // Add encrypted type
+            final EnumSet<Path.Type> type = EnumSet.copyOf(directory.getType());
+            type.add(Path.Type.encrypted);
+            type.remove(Path.Type.decrypted);
+            return new Path(intermediate, directoryIdHash.substring(2), type, attributes);
+        }
+        throw new NotfoundException(directory.getAbsolute());
+    }
+
+    protected byte[] toDirectoryId(final Session<?> session, final Path directory) throws BackgroundException {
+        if(new SimplePathPredicate(home).test(directory)) {
+            return ROOT_DIR_ID;
+        }
+        lock.readLock().lock();
+        try {
+            if(cache.contains(new SimplePathPredicate(directory))) {
+                return cache.get(new SimplePathPredicate(directory));
+            }
+        }
+        finally {
+            lock.readLock().unlock();
+        }
+        try {
+            log.debug("Acquire lock for {}", directory);
+            lock.writeLock().lock();
+            final byte[] id = this.load(session, directory);
+            cache.put(new SimplePathPredicate(directory), id);
+            return id;
+        }
+        finally {
+            lock.writeLock().unlock();
+        }
+    }
+
+    public void delete(final Path directory) {
+        lock.writeLock().lock();
+        try {
+            cache.remove(new SimplePathPredicate(directory));
+        }
+        finally {
+            lock.writeLock().unlock();
+        }
+    }
+
+    @Override
+    public void destroy() {
+        lock.writeLock().lock();
+        try {
+            cache.clear();
+        }
+        finally {
+            lock.writeLock().unlock();
+        }
+    }
+
+    @Override
+    public byte[] getOrCreateDirectoryId(final Session<?> session, final Path file) throws BackgroundException {
+        if(file.attributes().getDirectoryId() != null) {
+            return file.attributes().getDirectoryId();
+        }
+        final Path decrypted = file.getType().contains(AbstractPath.Type.encrypted) ? file.attributes().getDecrypted() : file;
+        return this.toDirectoryId(session, decrypted.getType().contains(AbstractPath.Type.file) ? decrypted.getParent() : decrypted);
+    }
+
+    @Override
+    public byte[] createDirectoryId(final Path directory) {
+        lock.writeLock().lock();
+        try {
+            final byte[] directoryId = random.random().getBytes(StandardCharsets.US_ASCII);
+            cache.put(new SimplePathPredicate(directory), directoryId);
+            return directoryId;
+        }
+        finally {
+            lock.writeLock().unlock();
+        }
     }
 
     protected byte[] load(final Session<?> session, final Path directory) throws BackgroundException {
